+++ components/common/FormField/FormField.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { View, Text, TextInput } from "react-native";
-import styles from "./styles";
 import { FormFieldProps } from "@/types/viewTrip/viewTrip";
+import styles from "./styles";
 
 export const FormField = ({
   label,
