+++ components/common/FormField/styles.tsx
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  inputGroup: {
+    marginBottom: 20,
+    zIndex: 1,
+  },
+
+  label: {
+    fontSize: 14,
+    color: "#666",
+    marginBottom: 8,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: "#ddd",
+    borderRadius: 8,
+    padding: 12,
+    fontSize: 16,
+  },
+});
+export default styles;
