+++ components/viewTrip/tripPlan/modal/editModal.tsx
@@ -3,13 +3,14 @@ import { Modal, View, ScrollView } from "react-native";
 import { useDebouncedSearch } from "@/hooks/map/useDebounceSearch";
 import styles from "./styles";
 import { SearchResult } from "@/types/map/map";
-import { categoryMap } from "@/constants/default";
 import { SearchResults } from "./SearchResults/SearchResults";
-import { FormField } from "./FormField";
+import { FormField } from "../../../common/FormField/FormField";
 import { ModalHeader } from "./ModalHeader";
 import { useLocationForm } from "@/hooks/viewTrip/useLocationForm";
 import { EditModalProps } from "@/types/viewTrip/viewTrip";
 import { ImagePickerSection } from "./ImagePickerSection";
+import { useEditSerachPlace } from "@/hooks/viewTrip/useEditSearchPlace";
+import FormGroup from "./FormGroup/FormGroup";
 
 const EditModal = ({
   visible,
@@ -25,32 +26,13 @@ const EditModal = ({
   const [isSearching, setIsSearching] = useState(false);
   const [searchResults, setSearchResults] = useState<SearchResult[]>([]);
   const [showSearchResults, setShowSearchResults] = useState(false);
-
-  const handleSearch = (text: string) => {
-    setFormData((prev) => ({ ...prev, name: text }));
-    if (text.length >= 2) {
-      setShowSearchResults(true);
-      useDebouncedSearch({
-        searchQuery: text,
-        setIsSearching,
-        setSearchResults,
-      });
-    } else {
-      setShowSearchResults(false);
-      setSearchResults([]);
-    }
-  };
-
-  const handleSelectPlace = (place: SearchResult) => {
-    setFormData((prev) => ({
-      ...prev,
-      name: place.place_name,
-      address: place.address_name,
-      category: categoryMap[place.category_group_code],
-    }));
-    setShowSearchResults(false);
-  };
-
+  const { handleSearch, handleSelectPlace } = useEditSerachPlace({
+    setFormData,
+    setShowSearchResults,
+    useDebouncedSearch,
+    setSearchResults,
+    setIsSearching,
+  });
   return (
     <Modal visible={visible} animationType="slide" onRequestClose={onClose}>
       <View style={styles.container}>
@@ -61,71 +43,15 @@ const EditModal = ({
         />
 
         <ScrollView style={styles.content}>
-          <ImagePickerSection
-            thumbnail={formData.thumbnail}
-            onImageSelect={(uri) =>
-              setFormData((prev) => ({ ...prev, thumbnail: uri }))
-            }
+          <FormGroup
+            formData={formData}
+            setFormData={setFormData}
+            handleSearch={handleSearch}
+            showSearchResults={showSearchResults}
+            isSearching={isSearching}
+            searchResults={searchResults}
+            handleSelectPlace={handleSelectPlace}
           />
-          <View style={styles.searchInputGroup}>
-            <FormField
-              label="Name"
-              value={formData.name}
-              onChangeText={handleSearch}
-              placeholder="장소를 검색해주세요..."
-            >
-              <SearchResults
-                show={showSearchResults}
-                isSearching={isSearching}
-                results={searchResults}
-                onSelect={handleSelectPlace}
-              />
-            </FormField>
-          </View>
-
-          <View style={styles.inputGroup}>
-            <FormField
-              label="Address"
-              value={formData.address}
-              onChangeText={(text) =>
-                setFormData((prev) => ({ ...prev, address: text }))
-              }
-              placeholder="Enter address"
-            />
-          </View>
-
-          <View style={styles.inputGroup}>
-            <FormField
-              label="Visit Time"
-              value={formData.visit_time}
-              onChangeText={(text) =>
-                setFormData((prev) => ({ ...prev, visitTime: text }))
-              }
-              placeholder="방문 시간을 입력해주세요..."
-            />
-          </View>
-
-          <View style={styles.inputGroup}>
-            <FormField
-              label="Category"
-              value={formData.category}
-              onChangeText={(text) =>
-                setFormData((prev) => ({ ...prev, category: text }))
-              }
-              placeholder="카테고리를 선택해주세요..."
-            />
-          </View>
-
-          <View style={styles.inputGroup}>
-            <FormField
-              label="Hashtag"
-              value={formData.hashtag}
-              onChangeText={(text) =>
-                setFormData((prev) => ({ ...prev, hashtag: text }))
-              }
-              placeholder="테그를 입력해주세요..."
-            />
-          </View>
         </ScrollView>
       </View>
     </Modal>
